comments

+++ src/homework7/Compress.java
@@ -6,6 +6,7 @@ import java.util.zip.DeflaterOutputStream;
 /**
  * Created by ns on 2/13/2016.
  */
+//[ZP] - 10 what about ZipOutputStream
 public class Compress {
 
     public static void main(String[] args){
@@ -17,10 +18,11 @@ public class Compress {
             FileOutputStream fout = new FileOutputStream(fileNameOut);
 
             DeflaterOutputStream out = new DeflaterOutputStream(fout);
+            
 
             int i;
             while((i=fin.read())!=-1){
-                out.write((byte)i);
+                out.write(i);
                 out.flush();
             }
